Make most fields private.
